Properly document the module search path.
Specifically, document --env-only, --list-all, --with-path,
PKG_CONFIG_LIBDIR, and PKG_CONFIG_PATH, all of which were
already mentioned, but very little was said about them,
and much of what *was* said was outdated.

This documents the intended behaviour of --with-path.
Our tree currently contains a (non-critical) bug related to it
that was fixed today in the upstream repository
and that we might wish to fix here, too, at our leisure.

This patch was accepted and pushed upstream by Ariadne Conill.