deprecate RW_SLEEPFAIL.
it was documented as: Wait for busy locks, but do not obtain them,
fail with EAGAIN instead.

nothing uses it anymore. the few things that did use it were too
tricky, so i'm not keen to leave it around for people to cause
confusion with in the future.

ok claudio@ mpi@ mvs@