Call in_pcbselsrc() and in6_pcbselsrc() with const sockaddr parameter.
Functions in_pcbselsrc() and in6_pcbselsrc() use the destination
sockaddr to determine a suitable source address.  As the destination
is only used for lookup, it should never be modified.  Use const
to let the compiler verify that.  On the way down the callstack,
also convert a bunch of other lookup functions to const.

OK kn@