Add initial regress test framework
Add a test framework for use in LibreSSL regression tests. This test
framework aims to be as lightweight and as simple to use as possible.

The design is mostly inspired by Go's test system, and aims to be a
drop-in utility in most existing regress tests.

ok jsing tb beck