add producer/consumer locking
these provide coordination between code that is producing or updating
data, and code that wants a consistent read of that data. it's set
up so that the producer can run without being blocked by consumers,
however, the consumer will have to retry reading the data if the
producer changed it while a consumer was running.

ok mvs@ claudio@