isakmpd: avoid undefined behavior in hash.c
Calling a function through an incompatible function pointer cast is UB.
Add wrapper functions for the hashing functions that pass the correct
member of the union of hash contexts to the libc hashing functions, so
we don't need to cast at all. This matches bin/mdt/md5.c r1.99.

ok mvs yasuoka